// src/types/HealthProfile.ts
export interface HealthProfile {
  fullName: string;
  dateOfBirth: string;
  gender: string;
  bloodGroup: string;
  phone: string;
  address: string;

  allergies: string;
  medicalConditions: string;
  medications: string;

  emergencyContactName: string;
  emergencyContactRelation: string;
  emergencyContactPhone: string;
}